rename signing -> common, refactor README sample SAS token append
