+++ sdk/storage/azure_storage_sas/src/common.rs
