+++ sdk/storage/azure_storage_sas/src/resource/queue.rs
@@ -22,8 +22,8 @@
 //! ```
 
 use crate::builder::SasBuilder;
-use crate::signing::sealed::Sealed;
-use crate::signing::{CommonFields, SasResource, ValidatedKey};
+use crate::common::sealed::Sealed;
+use crate::common::{CommonFields, SasResource, ValidatedKey};
 use crate::SAS_VERSION;
 
 /// A queue resource for user delegation SAS.
@@ -213,7 +213,7 @@ fn queue_udk_query_parameters(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::signing::test_support::{test_common, test_udk};
+    use crate::common::test_support::{test_common, test_udk};
     use time::macros::datetime;
 
     #[test]
